refactor: migrate project to monorepo
