+++ package.json
@@ -4,12 +4,9 @@
   "version": "1.3.0",
   "description": "",
   "scripts": {
-    "build:main": "rollup -c && cp dist/main.js activator.js",
-    "build:generator": "rollup -c rollup.generator.config.mjs && cp dist/generator.js generator.js",
-    "build": "npm run build:main && npm run build:generator",
-    "generate": "esno src/generator.ts",
-    "generate:file": "esno src/generator.ts --file",
-    "test": "vitest"
+    "build:core": "pnpm -C packages/core run build",
+    "build:generator": "pnpm -C packages/generator run build",
+    "build": "npm run build:core && npm run build:generator"
   },
   "keywords": [],
   "dependencies": {
@@ -21,13 +18,10 @@
     "@rollup/plugin-node-resolve": "^15.2.1",
     "@rollup/plugin-terser": "^0.4.3",
     "@types/node": "^20.6.0",
-    "@types/prompts": "^2.4.5",
     "esno": "^0.17.0",
-    "prompts": "^2.4.2",
     "rollup": "^3.29.1",
     "rollup-plugin-esbuild": "^5.0.0",
     "rollup-plugin-ts": "^3.4.5",
-    "toml": "^3.0.0",
-    "vitest": "^0.34.5"
+    "toml": "^3.0.0"
   }
 }
