+++ packages/core/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@as/core",
+  "version": "1.3.0",
+  "description": "",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "build": "rollup -c && mv ../../dist/main.js ../../dist/activator.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@as/shared": "workspace:^"
+  }
+}
