+++ packages/core/rollup.config.mjs
@@ -1,19 +1,18 @@
+import { defineConfig } from "rollup";
 import nodeResolve from "@rollup/plugin-node-resolve";
 import typescript from "rollup-plugin-ts";
-import { defineConfig } from "rollup";
 
 export default defineConfig({
   input: [
     "src/main.ts",
   ],
   output: {
-    dir: "dist",
+    dir: "../../dist",
     format: "cjs",
   },
   plugins: [
     nodeResolve(),
     typescript(),
-    // minify(),
   ],
   
 })
