+++ packages/dashboard/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@as/dashboard",
+  "version": "1.3.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
