+++ packages/generator/rollup.config.mjs
@@ -1,16 +1,17 @@
+import { defineConfig } from "rollup";
 import nodeResolve from "@rollup/plugin-node-resolve";
 import typescript from "rollup-plugin-ts";
-import { defineConfig } from "rollup";
 import commonjs from "@rollup/plugin-commonjs";
 import terser from "@rollup/plugin-terser";
 
 export default defineConfig({
   input: [
-    "src/generator.ts",
+    "./index.ts",
   ],
   output: {
-    dir: "dist",
+    dir: "../../dist",
     format: "cjs",
+    name: "generator",
   },
   plugins: [
     nodeResolve(),
