+++ packages/shared/src/modules/elpass/activate.ts
