+++ packages/shared/src/modules/gumroad/alogrithm/index.ts
