+++ packages/shared/src/modules/gumroad/validate.ts
