+++ packages/shared/src/modules/lemon-squeezy/activate.ts
