+++ packages/shared/src/modules/lemon-squeezy/validate.ts
