+++ packages/shared/src/modules/raycast/activate.ts
