+++ packages/shared/src/modules/raycast/customs/ai/completions/gemini.ts
