+++ packages/shared/src/modules/raycast/customs/ai/completions/index.ts
