+++ packages/shared/src/modules/raycast/customs/ai/completions/openai.ts
